Refines Kepler logic a tad - adds kepler logic unit tests

+++ src/typedefs.h
@@ -4,4 +4,4 @@
 typedef unsigned int entityID;
 
 // Seconds relative to the J2000 Epoch
-typedef unsigned long int universeTime;
+typedef double universeTime;
